tests/test_main.py
## IMPORTS
import unittest
from unittest.mock import patch, MagicMock
from fastapi.testclient import TestClient
from sqlalchemy.orm import sessionmaker

## SETUP TEST ENVIRONMENT
# Mock settings to avoid importing actual settings which might not be available in the test environment
mock_settings = MagicMock()
mock_settings.PROJECT_NAME = "Test Project"
mock_settings.PROJECT_VERSION = "1.0"

# Mock database session to avoid actual DB calls during tests
mock_session_local = MagicMock(spec=sessionmaker)

# Mock engine to avoid actual DB engine calls
mock_engine = MagicMock()

## TEST CASES

## <CREATE_APP_TEST>
class TestCreateApp(unittest.TestCase):
    @patch("main.settings", mock_settings)
    @patch("main.SessionLocal", mock_session_local)
    @patch("main.engine", mock_engine)
    def test_create_app(self):
        """
        Test the creation of the FastAPI app with all its routes and configurations.
        """
        from main import create_app  # Import here to use patched values

        app = create_app()
        client = TestClient(app)

        # Test if the app is an instance of FastAPI
        self.assertIsInstance(app, FastAPI)

        # Test if all expected routes are included
        routes = [route.path for route in app.routes]
        expected_routes = [
            "/api/v1/agents",
            "/api/v1/roles",
            "/api/v1/influences",
            "/api/v1/stages",
            "/api/v1/groups",
            "/api/v1/tasks",
            "/api/v1/news",
            "/api/v1/recommendations",
            "/api/v1/training",
            "/api/v1/feedback",
            "/api/v1/nlp"  # Note: '/api/v1/nlp' is expected to be present multiple times due to different endpoints
        ]
        for expected_route in expected_routes:
            self.assertIn(expected_route, routes)

        # Test if the app title and version match the settings
        self.assertEqual(app.title, mock_settings.PROJECT_NAME)
        self.assertEqual(app.version, mock_settings.PROJECT_VERSION)

## <DATABASE_DEPENDENCY_TEST>
class TestDatabaseDependency(unittest.TestCase):
    @patch("main.SessionLocal", mock_session_local)
    @patch("main.engine", mock_engine)
    def test_get_db_dependency(self):
        """
        Test the get_db dependency for yielding a database session and closing it properly.
        """
        from main import create_app  # Import here to use patched values
        app = create_app()
        with TestClient(app) as client:
            with patch("main.get_db") as mock_get_db:
                mock_get_db.return_value = mock_session_local()
                response = client.get("/api/v1/agents")  # Using agents as an example endpoint
                self.assertTrue(mock_get_db.called)
                self.assertEqual(response.status_code, 200)

## <ERROR_HANDLERS_TEST>
class TestErrorHandlers(unittest.TestCase):
    @patch("main.settings", mock_settings)
    @patch("main.SessionLocal", mock_session_local)
    @patch("main.engine", mock_engine)
    def test_error_handlers(self):
        """
        Test if custom error handlers are correctly set up and handling errors as expected.
        """
        from main import create_app  # Import here to use patched values
        app = create_app()
        client = TestClient(app)

        # Assuming there's a custom error handler for 404 errors
        response = client.get("/api/v1/nonexistent_endpoint")
        self.assertEqual(response.status_code, 404)
        # Further assertions can be made based on the custom error response structure

## RUNNING TESTS
if __name__ == "__main__":
    unittest.main()
